+++ packages/apps/dev-wallet/src/Components/Sidebar/style.css.ts
@@ -42,7 +42,6 @@ export const sidebarLinkClass = style([
     padding: 'md',
   }),
   {
-    color: sidebarTextColorClass,
     minWidth: '192px',
     selectors: {
       '&.active': {