+++ packages/apps/docs/src/components/DocsCard/styles.css.ts
@@ -29,7 +29,7 @@ export const cardClass = style([
   }),
   {
     transition: 'all .3s ease',
-    backgroundColor: tokens.kda.foundation.color.palette.blue.n1,
+    backgroundColor: tokens.kda.foundation.color.neutral.n1,
     color: tokens.kda.foundation.color.text.base.default,
   },
 ]);
@@ -37,7 +37,7 @@ export const cardClass = style([
 export const docsCardLink = style([
   atoms({
     textDecoration: 'none',
-    fontWeight: 'secondaryFont.bold',
+    fontWeight: 'primaryFont.bold',
   }),
   {
     color: tokens.kda.foundation.color.text.base.default,