+++ packages/libs/kode-ui/src/components/Stepper/index.ts
@@ -1,0 +1,4 @@
+export { Step } from './Step';
+export type { IStepProps } from './Step';
+export { Stepper } from './Stepper';
+export type { IStepperProps } from './Stepper';